requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
astropy==5.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dateparser==1.1.5
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pycodestyle==2.10.0
pyerfa==2.0.0.1
pyparsing==3.0.9
pyregion==2.2.0
PySide2==5.15.2.1
PySimpleGUI==4.60.4
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
shiboken2==5.15.2.1
six==1.16.0
SQLAlchemy==1.4.46
TA-Lib @ file:///D:/mineFlask/TA_Lib-0.4.24-cp39-cp39-win_amd64.whl
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.13
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
aiohttp==3.8.3
aiosignal==1.3.1
astropy==5.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dateparser==1.1.5
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pycodestyle==2.10.0
pyerfa==2.0.0.1
pyparsing==3.0.9
pyregion==2.2.0
PySide2==5.15.2.1
PySimpleGUI==4.60.4
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
shiboken2==5.15.2.1
six==1.16.0
SQLAlchemy==1.4.46
TA-Lib @ file:///D:/mineFlask/TA_Lib-0.4.24-cp39-cp39-win_amd64.whl
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.13
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
